refactor(test): Arrange mock codes
